+++ lib/blocs/product_bloc/product_state.dart
