+++ lib/model/models.dart
@@ -1,2 +1,4 @@
-export 'product.dart';
-export '../user/user.dart';
+export 'product/product.dart';
+export 'user/user_data.dart';
+export 'user/user_meta.dart';
+export 'user/user_repository.dart';
