+++ lib/model/product/product.dart
