+++ lib/screens/product/product.dart
@@ -1,0 +1,2 @@
+export 'product_main.dart';
+export 'product_screen.dart';
